+++ core/src/transport/dummy.rs
@@ -79,10 +79,6 @@ impl<TOut> Transport for DummyTransport<TOut> {
         Err(TransportError::MultiaddrNotSupported(addr))
     }
 
-    fn address_translation(&self, _server: &Multiaddr, _observed: &Multiaddr) -> Option<Multiaddr> {
-        None
-    }
-
     fn poll(
         self: Pin<&mut Self>,
         _: &mut Context<'_>,