+++ protocols/autonat/src/v2/server.rs
@@ -1,5 +1,0 @@
-mod behaviour;
-mod handler;
-
-pub use behaviour::Behaviour;
-pub use behaviour::Event;